Refactor: "obs_document.ObsVault" is now "obs_vault.ObsVault".
Keeps each major class in its own file; seems cleaner.
